+++ src/main/kotlin/pl/umk/mat/locals/repositories/OfferRepository.kt
@@ -11,7 +11,7 @@ import pl.umk.mat.locals.models.Tag
 interface OfferRepository : CrudRepository<Offer, Long> {
     @Query(value = "SELECT * FROM offer WHERE (ST_Distance_Sphere( point(lon, lat), point(?2, ?1) ) / 1000) <= ?3",
             nativeQuery = true)
-    fun findAllOffersByPoint(lat: Float, lon: Float, radius: Long): List<Offer>
+    fun findAllOffersByPoint(lat: Double, lon: Double, radius: Long): List<Offer>
 
     fun findAllOffersByCity(city: String): List<Offer>
     @Query(value = "SELECT * FROM offer ORDER BY RAND() LIMIT 10",
@@ -19,5 +19,5 @@ interface OfferRepository : CrudRepository<Offer, Long> {
     fun findRandomOffers(): List<Offer>
 
     fun findAllByTagsIn(list: List<Tag>): List<Offer>
-    //fun findAllOffersByTags(list: List<Tag>): List<Offer>
+
 }
